Factor tt::is_std_array into its own file
