+++ src/Utilities/PrettyType.hpp
@@ -24,9 +24,9 @@
 
 #include "Utilities/Requires.hpp"
 #include "Utilities/TMPL.hpp"
-#include "Utilities/TypeTraits.hpp"
 #include "Utilities/TypeTraits/ArraySize.hpp"
 #include "Utilities/TypeTraits/IsA.hpp"
+#include "Utilities/TypeTraits/IsStdArray.hpp"
 
 /// \cond
 #define PRETTY_TYPE_USE_BOOST
