+++ tests/Unit/Pypp/CheckWithRandomValues.hpp
@@ -55,6 +55,7 @@
 #include "Utilities/TypeTraits.hpp"
 #include "Utilities/TypeTraits/FunctionInfo.hpp"
 #include "Utilities/TypeTraits/IsA.hpp"
+#include "Utilities/TypeTraits/IsStdArray.hpp"
 #include "tests/Unit/Pypp/Pypp.hpp"
 #include "tests/Unit/TestHelpers.hpp"
 #include "tests/Utilities/MakeWithRandomValues.hpp"
