+++ tests/Unit/Pypp/Pypp.hpp
@@ -14,6 +14,7 @@
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TypeTraits.hpp"
 #include "Utilities/TypeTraits/IsA.hpp"
+#include "Utilities/TypeTraits/IsStdArray.hpp"
 #include "tests/Unit/Pypp/PyppFundamentals.hpp"
 
 /// \ingroup TestingFrameworkGroup
